update to pull in VM trait from external location
