+++ Cargo.toml
@@ -25,7 +25,17 @@ fvm_ipld_car = { version = "0.7.0"}
 fvm_ipld_encoding = { version = "0.4.0" }
 fvm_ipld_hamt = { version = "0.7.0" }
 fvm_shared = { version = "~3.4.0", default-features = false }
+vm_api = { version = "1.0.0", git="https://github.com/filecoin-project/builtin-actors", branch="alexytsu/export-test-vm" }
 
+anyhow = { version = "~1.0.47" }
+blake2b_simd = { version = "1.0" }
+bls-signatures = { version = "0.13", default-features = false }
+futures = { version = "~0.3.19" }
+libsecp256k1 = { version = "0.7.1"}
+multihash = { version = "0.18.1", default-features = false }
+num-traits = "0.2.14"
+rand_chacha = "~0.3"
+replace_with = "0.1.7"
 
 [patch.crates-io]
 # fvm = {git="https://github.com/helix-onchain/ref-fvm", branch="master"}
