+++ builtin/Cargo.toml
@@ -26,7 +26,6 @@ fil_actor_system = { version = "12.0.0", git = "https://github.com/filecoin-proj
 fil_actor_verifreg = { version = "12.0.0", git = "https://github.com/filecoin-project/builtin-actors", branch = "master", features = [] }
 fil_actors_runtime = { version = "12.0.0", git = "https://github.com/filecoin-project/builtin-actors", branch = "master", features = [] }
 
-
 fvm = { workspace = true }
 fvm_actor_utils = { workspace = true } 
 fvm_ipld_bitfield = { workspace = true }
@@ -34,19 +33,18 @@ fvm_ipld_blockstore = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 fvm_ipld_hamt = { workspace = true } 
 fvm_shared = { workspace = true }
+vm_api = { workspace = true }
 
 # Used directly in some tests
 frc46_token = { workspace = true }
 
-anyhow = "~1.0.47"
-blake2b_simd = "1.0.0"
-bls-signatures = { version = "0.13", default-features = false }
+anyhow = { workspace = true }
+blake2b_simd = { workspace = true }
+bls-signatures = { workspace = true }
 cid = { workspace = true }
-libsecp256k1 = { version = "0.7" }
-multihash = { version = "~0.16.1", default-features = false }
-rand_chacha = "~0.3"
-thiserror = "~1.0.30"
-num-traits = "0.2.14"
+libsecp256k1 = { workspace = true }
+rand_chacha = { workspace = true }
+num-traits = { workspace = true }
 
 [dev-dependencies]
 log = "0.4"
