+++ builtin/tests/util/hookup.rs
@@ -7,11 +7,13 @@ use fvm_shared::state::StateTreeVersion;
 use fvm_shared::version::NetworkVersion;
 use fvm_shared::METHOD_SEND;
 // use fvm_workbench_api::analysis::TraceAnalysis;
-use fvm_workbench_api::wrangler::{ExecutionWrangler, VM};
+use fvm_workbench_api::wrangler::ExecutionWrangler;
 use fvm_workbench_api::WorkbenchBuilder;
 use fvm_workbench_builtin_actors::genesis::{create_genesis_actors, GenesisSpec};
 use fvm_workbench_vm::builder::FvmBenchBuilder;
 use fvm_workbench_vm::externs::FakeExterns;
+use fvm_workbench_vm::primitives::FakePrimitives;
+use vm_api::VM;
 
 #[test]
 fn test_hookup() {
@@ -28,7 +30,8 @@ fn test_hookup() {
     let spec = GenesisSpec::default(manifest_data_cid);
     let genesis = create_genesis_actors(&mut builder, &spec).unwrap();
     let bench = builder.build().unwrap();
-    let wrangler = ExecutionWrangler::new_default(bench, Box::new(store));
+    let wrangler =
+        ExecutionWrangler::new_default(bench, Box::new(store), Box::new(FakePrimitives {}));
 
     let result = wrangler
         .execute_message(
