+++ vm/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod bench;
 pub mod builder;
 pub mod externs;
+pub mod primitives;
